Clean up specs
The new design of the formatter and its collaborators did not require as
extensive of testing
