+++ lib/flatware/step_result.rb
@@ -1,19 +1,4 @@
 module Flatware
-  class ScenarioResult
-    attr_reader :status
-    def initialize(status)
-      @status = status
-    end
-
-    def passed?
-      status == :passed
-    end
-
-    def failed?
-      status == :failed
-    end
-  end
-
   class StepResult
     attr_reader :status, :exception
